Use custom time for persisting action config
